pyproject.toml
[project]
name = "poop-tracker-bot"
version = "0.1.0"
description = "Telegram bot for tracking bowel movements for people with Crohn's disease and ulcerative colitis"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dynamic = ["dependencies", "optional-dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
aiogram = "3.25.0"
python-dotenv = "1.0.0"
sqlalchemy = "2.0.30"
greenlet = "3.0.3"
asyncpg = "0.30.0"
aiosqlite = "0.20.0"
alembic = "1.14.1"
pytz = "2024.2"
python-dateutil = "2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
pytest-asyncio = "0.23.7"
black = "24.10.0"
isort = "5.13.2"
mypy = "1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
